Move /outputs and /balance apis from wallet group
